+++ src/screens/EmployerScreen.js
@@ -1,4 +1,5 @@
-import React, { useEffect, useState } from 'react'
+import React, { useEffect, useState, useContext } from 'react'
+import { AuthContext } from '../context/auth'
 import axios from 'axios'
 import moment from 'moment'
 import { useNavigate } from 'react-router-dom'
@@ -12,6 +13,8 @@ const EmployerScreen = () => {
   const [refresh, setRefresh] = useState('false')
 
   const history = useNavigate()
+
+  const {user} = useContext(AuthContext)
 
   const token = localStorage.getItem('jwtToken')
   if(!token) {
@@ -35,11 +38,14 @@ const EmployerScreen = () => {
 
   const deletHandler = async (id) => {
     if(window.confirm('Are you sure you want to delete this Employer?')){
-      const deleteEmpl = await axios.delete(`/api/employers/profile/${id}`, config)
-      if(deleteEmpl){
-        console.log(deleteEmpl)
+      if(user && user.email === process.env.REACT_APP_DEMO_MAIL){
+        window.alert('This employer gets deleted here! This is just a demo! No new emlpoyer deleted!')
+      } else {
+        const deleteEmpl = await axios.delete(`/api/employers/profile/${id}`, config)
+        if(deleteEmpl){
         setRefresh(true)
-      }
+        }
+      } 
     }
   }
 
@@ -83,7 +89,7 @@ const EmployerScreen = () => {
                       <td>{employer.location.city}</td>
                       <td>{employer.location.state}</td>
                       <td>{employer.location.country}</td>
-                      <td>{employer.applications.length}</td>
+                      <td>{employer.applicationCount}</td>
                       <td>{moment(employer.createdAt).format('MM/DD/YY')}</td>
                       <td>
                         <LinkContainer to={`/employers/profile/${employer._id}`}>
